restructure (adds system.rs)
